Sudoku.jl Documentation

Overview

Below are two are two Sudoku solvers: One takes a puzzle as a matrix while the other takes a file. In addition, a few utility functions that help in the process are listed.

get_block : This function retrieves sub-blocks of a Sudoku matrix.

get_blk_idx : This function determines which sub-block matrix to pick.

has_dups : Checks a given vector/matrix for non-zero duplicates.

consist_chk : Checks for consistency of puzzle, partial, or full Sudoku solution matrix.

check_sudoku_solution: Checks a proposed solution matrix against its puzzle matrix.

solve_sudoku : Solves a Sudoku puzzle represented as a matrix.

solve_sudoku_puzzle : Solves a Sudoku puzzle represented as a file.

Utilities

Sudoku.get_block — Function

get_block(S, i, j)

Get the (i, j) sub-block matrix of S.

Arguments

S::Matrix{Int8} – A sudoku matrix.

i::Int64 – The $i^{i\rm th}$ block row entry.

j::Int64 – The $j^{j\rm th}$ block column entry.

Return

::Matrix{Int8} – The $(i^{\rm th}, j^{\rm th})$ sub-block matrix.

Sudoku.get_blk_idx — Function

get_blk_idx(h)

Get the sub-block index pair for a given index pair of a Sudoku matrix.

Arguments

h::CartesianIndex – The index into a Sudoku matrix.

Examples

The code below gets the sub-block index pair, (2,2) (the index representing the middle 3x3 block matrix), when passed the Sudoku matrix index pair (4,6).

julia> get_blk_idx(CaresianIndex(4, 6))
(2,2)

Return

::Tuple{Int64, Int64} – The index pair of the sub-matrix block.

Sudoku.has_dups — Function

has_dups(v)

Check if a vector/matrix has duplicate numeric (other than 0) entries.

Arguments

v :: Union{Vector{Int8}, Matrix{Int8}}

Return

::Bool – If true there exists at least one duplicate.

Sudoku.consist_chk — Function

consist_chk(S)

Checks the consistency of a Sudoku matrix, S.

This means that we check that there are no (non-zero) duplicate entries in any rows, columns, or sub-blocks.

Argumens

S::Matrix{Int8} – A Sudoku puzzle, proposed solution, or intermediate solution.

Return

::Bool – Returns true if Sudoku matrix is consistent.

Sudoku.check_sudoku_solution — Function

check_sudoku_solution(SP, SS)

Checks that SS is a solution of the puzzle, SP.

It does this by doing the following:

Checks for any (non-zero) duplicate entries in any rows, columns, or sub-blocks.

Checks that the row, column, and sub-block sums are SUDOKU_SUM.

Checks that the proposed solution is consistent with the puzzle matrix, SP.

Arguments

SP::Matrix{Int8} – A Sudoku puzzle in matrix form. Zeros represent blanks.

SS::Matrix{Int8} – Proposed solution for SP.

Returns

::Bool – If true, the proposed solution is correct.

Sudoku Solvers

Sudoku.solve_sudoku — Function

solve_sudoku(SP, rec_count; verbose=false)

Helper function that does the work of the top level solver.

Arguments

S::Matrix{Int8} – A Sudoku puzzle matrix.

rec_count::Int64 – The count of the number of times this function has been called.

Keyword Arguments

verbose::Bool=false – If true, print out extra information.

Return

(ok, SS)

ok::Bool – If true, a proposed solution was found.

S::Matrix{Int8}} – A proposed, or inconsistent solution matrix.

solve_sudoku(S; verbose=false)

Solves a Sudoku puzzle represented as a matrix.

The value, 0, is used in a puzzle matrix to represent a blank.

Arguments

S::Matrix{Int8} – A Sudoku puzzle matrix.

Keyword Arguments

verbose::Bool=false – If true, print out extra information.

Return

(ok, chk_sol, SS)

ok::Bool – If true, a proposed solution was found.

chk_sol::Bool – If true, the proposed solution is correct.

SS::Matrix{Int8}} – A proposed, or inconsistent/incomplete solution matrix.

Sudoku.solve_sudoku_file — Function

solve_sudoku_file(puzzle_file_name; <keyword arguments>])

Solves and prints a solution of a Sudou puzzle file.

The file is the name of a CSV file (without extension). The file format is 9 rows of values, 0-9, with "0" representing a blank.

NOTE: The CSV file should not have a header.

Arguments

puzzle_file_name::AbstractString – The puzzle file name without the extension.

Keyword Arguments

puzzle_dir=joinpath(@__DIR__, "../puzzles") :: AbstractString – The path to the puzzle file directory.

verbose=false :: Bool – If true, print more output.

Return

::Nothing

Index

Sudoku.check_sudoku_solution

Sudoku.consist_chk

Sudoku.get_blk_idx

Sudoku.get_block

Sudoku.has_dups

Sudoku.solve_sudoku

Sudoku.solve_sudoku_file